address comment

+++ meta/autoid/autoid.go
@@ -262,7 +262,7 @@ func (alloc *allocator) alloc4Signed(tableID int64) (int64, error) {
 	}
 
 	alloc.base++
-	log.Debugf("[kv] Alloc id %d, table ID:%d, from %p, database ID:%d", uint64(alloc.base), tableID, alloc, alloc.dbID)
+	log.Debugf("[kv] Alloc id %d, table ID:%d, from %p, database ID:%d", alloc.base, tableID, alloc, alloc.dbID)
 	return alloc.base, nil
 }
